fix: Stripe minimum top-up amount validation
The TOS states: "a user has to deposit exactly their total negative
balance (such that their balance becomes 0 euros after the deposit)
or at least 10 euros". However, the existing logic allowed top-ups
that would result in the user still have a negative balance.

For example, if the user has a balance of -€15,00 and tries to top
up with €10,00 this is allowed because

```
request.amount.amount >= MIN_TOPUP
```

returns `true`, while the TOS states this can only be possible if
the user's balance is non-negative.